move image collection files into subdirectory
