+++ src/Components/Publishing/Sections/ImageCollection/ImageCollection.tsx
@@ -2,10 +2,10 @@ import { find } from "lodash"
 import React from "react"
 import sizeMe from "react-sizeme"
 import styled from "styled-components"
-import fillwidthDimensions from "../../../Utils/fillwidth"
-import { pMedia } from "../../Helpers"
-import { SIZE_ME_REFRESH_RATE } from "../Constants"
-import { SectionLayout } from "../Typings"
+import fillwidthDimensions from "../../../../Utils/fillwidth"
+import { pMedia } from "../../../Helpers"
+import { SIZE_ME_REFRESH_RATE } from "../../Constants"
+import { SectionLayout } from "../../Typings"
 import { Artwork } from "./Artwork"
 import { Image } from "./Image"
 import { ImageCollectionItem } from "./ImageCollectionItem"
