+++ src/Components/Publishing/Sections/ImageCollection/ImageCollectionItem.tsx
@@ -1,6 +1,6 @@
 import React from "react"
 import styled, { StyledFunction } from "styled-components"
-import { pMedia } from "../../Helpers"
+import { pMedia } from "../../../Helpers"
 
 interface FillwidthItemProps extends React.HTMLProps<HTMLDivElement> {
   margin?: number
