+++ src/Components/Publishing/Sections/ImageCollection/__test__/Image.test.tsx
@@ -1,7 +1,7 @@
 import "jest-styled-components"
 import React from "react"
 import renderer from "react-test-renderer"
-import { Images } from "../../Fixtures/Components"
+import { Images } from "../../../Fixtures/Components"
 import { Image } from "../Image"
 
 it("renders properly", () => {
@@ -17,11 +17,11 @@ it("renders a long caption properly", () => {
 it("renders a react child as caption properly", () => {
   const image = renderer
     .create(
-      <Image image={Images[2]}>
-        <div>
-          <p>A React child as caption.</p>
-        </div>
-      </Image>
+    <Image image={Images[2]}>
+      <div>
+        <p>A React child as caption.</p>
+      </div>
+    </Image>
     )
     .toJSON()
   expect(image).toMatchSnapshot()
