+++ src/Components/Publishing/Sections/ImageSetPreview.tsx
@@ -5,7 +5,6 @@ import { resize } from "../../../Utils/resizer"
 import { pMedia } from "../../Helpers"
 import { Fonts } from "../Fonts"
 import { IconImageSet } from "../Icon/IconImageSet"
-// import { ImageWrapper } from "./ImageWrapper"
 
 type Layout = "mini" | "full"
 
@@ -49,12 +48,7 @@ export class ImageSetPreview extends React.PureComponent<Props, null> {
       width,
       alt: this.props.section.title || "Open Slideshow"
     }
-
     return <img {...imageProps} />
-    // return (
-    //   <ImageWrapper {...imageProps} />
-    // )
-    // return <img src={src} width={width} height={height} alt={this.props.section.title || "Open Slideshow"} />
   }
 
   onClick = () => {
@@ -210,221 +204,3 @@ const SubTitleCount = styled.div`
     ${Fonts.unica("s12")}
   `}
 `
-
-// import PropTypes from "prop-types"
-// import React, { Component } from "react"
-// import styled, { StyledFunction } from "styled-components"
-// import { resize } from "../../../Utils/resizer"
-// import { pMedia } from "../../Helpers"
-// import { Fonts } from "../Fonts"
-// import { IconImageSet } from "../Icon/IconImageSet"
-// import { ImageWrapper } from "./ImageWrapper"
-
-// type Layout = "mini" | "full"
-
-// interface DivLayoutProps {
-//   layout?: Layout
-// }
-
-// export interface Props {
-//   section: {
-//     type: string
-//     images: Array<{
-//       url?: string
-//       image?: string
-//       index?: any
-//     }>
-//     layout?: Layout
-//     title?: string
-//   }
-// }
-
-// export class ImageSetPreview extends Component<Props, null> {
-
-//   static contextTypes = {
-//     onViewFullscreen: PropTypes.func,
-//   }
-
-//   onClick = () => {
-//     if (this.context.onViewFullscreen) {
-//       this.context.onViewFullscreen(this.props.section.images[0].index)
-//     }
-//   }
-
-//   textSection() {
-//     let {
-//       section: {
-//         images,
-//         layout,
-//         title
-//       }
-//     } = this.props
-
-//     if (!title) {
-//       title = images.length + " Images"
-//     }
-
-//     return (
-//       <TitleWrapper layout={layout}>
-//         <Title>
-//           {title}
-//         </Title>
-//         <SubTitle>
-//           <SubTitlePrompt>View Slideshow</SubTitlePrompt>
-//           { title &&
-//             <SubTitleCount>
-//               {images.length} Images
-//             </SubTitleCount>
-//           }
-//         </SubTitle>
-//       </TitleWrapper>
-//     )
-//   }
-
-//   render() {
-//     const {
-//       section: {
-//         images,
-//         layout,
-//         title
-//       }
-//     } = this.props
-
-//     const isFullLayout = layout === 'full'
-//     const image = images[0]
-//     const src = resize(image.url || image.image, { width: 1200 })
-//     const imageProps = {
-//       src,
-//       width: "auto",
-//       height: "100%"
-//     }
-
-//     if (isFullLayout) {
-//       imageProps.width = "100%"
-//       imageProps.height = "auto"
-//     }
-
-//     return (
-//       <div style={{ position: "relative", width: "100%" }}>
-//         {(() => {
-//           if (isFullLayout) {
-//             return (
-//               <FullWrapper onClick={this.onClick}>
-//                 {this.textSection()}
-
-//                 <IconContainer>
-//                   <IconImageSet />
-//                 </IconContainer>
-//               </FullWrapper>
-//             )
-//           } else {
-//             return (
-//               <MiniWrapper onClick={this.onClick}>
-//                  <ImageWrapper {...imageProps}
-//                   alt={title || "Open Slideshow"}
-//                 />
-
-//                 <MiniInner>
-//                   {this.textSection()}
-
-//                   <IconContainer>
-//                     <IconImageSet />
-//                   </IconContainer>
-//                 </MiniInner>
-//               </MiniWrapper>
-//             )
-//           }
-//         })()}
-
-//         <ImageWrapper {...imageProps}
-//           alt={title || "Open Slideshow"}
-//         />
-//       </div>
-//     )
-//   }
-// }
-
-// const div: StyledFunction<DivLayoutProps & React.HTMLProps<HTMLDivElement>> = styled.div
-
-// const IconContainer = styled.div`
-//   height: 45px;
-//   position: relative;
-//   margin-left: 40px;
-//   text-align: right;
-//   > svg {
-//     height: 98%;
-//   }
-//   ${pMedia.xs`
-//     display: none;
-//   `}
-// `
-
-// const FullWrapper = styled.div`
-//   position: absolute;
-//   bottom: 20px;
-//   left: 20px;
-//   min-height: 50px;
-//   width: auto;
-//   max-width: calc(100% - 80px);
-//   display: flex;
-//   justify-content: space-between;
-//   align-items: center;
-//   border-radius: 2px;
-//   background: rgba(255, 255, 255, 0.8);
-//   box-shadow: 0 0 10px 0 rgba(0, 0, 0, 0.3);
-//   padding: 20px;
-//   cursor: pointer;
-//   z-index: 1;
-//   &:hover {
-//     background: rgba(0, 0, 0, 0.6);
-//     color: white;
-//     ${IconContainer} {
-//       fill: white;
-//     }
-//   }
-// `
-// const TitleWrapper = div`
-//   display: flex;
-//   flex-direction: column;
-//   justify-content: space-between;
-//   height: ${props => (props.layout === "full" ? "50px" : "100%")};
-// `
-// const MiniWrapper = styled.div`
-//   height: 100px;
-//   display: flex;
-//   align-items: center;
-//   justify-content: flex-start;
-//   padding: 10px 20px 10px 10px;
-//   border: 1px solid #e5e5e5;
-//   cursor: pointer;
-// `
-// const MiniInner = styled.div`
-//   display: flex;
-//   justify-content: space-between;
-//   flex: 1;
-//   width: auto;
-//   margin-left: 20px;
-// `
-// const Title = styled.div`
-//   ${Fonts.unica("s19", "medium")}
-//   margin-bottom: 8px;
-//   line-height: 1.1em;
-//   ${pMedia.xs`
-//     ${Fonts.unica("s16", "medium")}
-//   `}
-// `
-// const SubTitle = styled.div`
-//   display: flex;
-// `
-// const SubTitlePrompt = styled.div`
-//   ${Fonts.unica("s14", "medium")}
-//   ${pMedia.xs`
-//     ${Fonts.unica("s12", "medium")}
-//   `}
-// `
-// const SubTitleCount = styled.div`
-//   ${Fonts.unica("s14")} margin-left: 20px;
-//   ${pMedia.xs`
-//     ${Fonts.unica("s12")}
-//   `}
-// `
